+++ src/parsers/problem/CodeRunProblemParser.ts
@@ -5,7 +5,11 @@ import { Parser } from '../Parser';
 
 export class CodeRunProblemParser extends Parser {
   public getMatchPatterns(): string[] {
-    return ['https://coderun.yandex.ru/problem/*', 'https://coderun.yandex.ru/*/problems/*'];
+    return [
+      'https://coderun.yandex.ru/problem/*',
+      'https://coderun.yandex.ru/*/problems/*',
+      'https://coderun.yandex.ru/*/problem/*',
+    ];
   }
 
   public async parse(url: string, html: string): Promise<Sendable> {
@@ -14,9 +18,10 @@ export class CodeRunProblemParser extends Parser {
 
     task.setName(elem.querySelector('h1').textContent);
 
-    const limitsElems = elem.querySelectorAll('dl[class^="Description_description-runtime-limits"] > dd > p');
-    task.setTimeLimit(parseInt(limitsElems[0].textContent.split(' ')[0]) * 1000);
-    task.setMemoryLimit(parseInt(limitsElems[1].textContent.split(' ')[0]));
+    const limitsElems = elem.querySelectorAll('dl[class^="Description_description-runtime-limits"] > dd');
+    const [timeValue, timeUnit] = limitsElems[0].textContent.trim().split(/\s+/);
+    task.setTimeLimit(parseInt(timeValue) * (timeUnit === 'мс' ? 1 : 1000));
+    task.setMemoryLimit(parseInt(limitsElems[1].textContent.trim().split(/\s+/)[0]));
 
     elem.querySelectorAll('.io-sample').forEach(sampleElem => {
       const input = sampleElem.querySelector('.input-snippet code').textContent;